add Makefile and do server build in docker
devs wanting to speed up the build can just
comment the gradle build line in the Dockerfile
and do the build manually on the host, then
`the COPY . .` line will move the cached build
from the host into the image.

I tried for a long time to try to get the
docker build to do the caching for us, but,
no dice.