Docker support (#158)
* add docker-compose and Dockerfile for server

* add Dockerfile for client and add to docker-compose.yml

* precompile client in dockerfile to avoid on-the-fly middleware compilation

* add Makefile and do server build in docker

devs wanting to speed up the build can just
comment the gradle build line in the Dockerfile
and do the build manually on the host, then
`the COPY . .` line will move the cached build
from the host into the image.

I tried for a long time to try to get the
docker build to do the caching for us, but,
no dice.

* add make run-rebuild to rebuild docker images

* create seed.dat

* don't use docker-compose secret, just volume, we need write access to create the seed

* pass url and password on command line

* add documentation for make run